Update to use virtual desktop notification listener
Right now the desktop index is still calculated every rainmeter update.
In the future I may refactor again so that it can be calculated at
desktop change and with that change the constructor will go back to
needing 0 arguments. CPU usage is better than before and is still very
good (basically 0% even when running debug builds). Other than a few C++
limitations I dislike that would require a lot of overhauling it is
pretty much done.